update: scoreboard style update
