+++ src/components/scoreboard/scoreboard-column/scoreboard-column.style.scss
@@ -8,6 +8,7 @@
   margin: 0 0.2rem;
   padding: 1rem 0.2rem ;
   width: 100%;
+  max-width: 10rem;
 
   &__player{
     color:green;
