+++ src/components/scoreboard/scoreboard.style.scss
@@ -3,4 +3,5 @@
   display: flex;
   justify-content: space-between;
   width: 100%;
+  max-width: 25rem;
 }
