+++ src/sass/_global.scss
@@ -8,9 +8,15 @@ body, #root{
   justify-content: space-between;
   min-height: 100vh;
   width: 100%;
+}
 
-  &#root{
-    box-sizing: border-box;
-    padding: 5rem;
+#root{
+  box-sizing: border-box;
+  padding: 5rem;
+}
+
+@media screen and (max-width:700px) {
+  #root{
+    padding:1rem;
   }
 }
